v1.0.0/requirements.txt
six
python-slugify